modified:   2-printf.c 	modified:   main.h 	modified:   printf.c

+++ main.h
@@ -6,5 +6,5 @@ int continue_printf(char, va_list, int);
 int print_char(char, int);
 int print_string(char *, int);
 int print_num(int, int);
-int print_binary(unsigned int, int count);
+int print_binary(unsigned long int, int count);
 #endif
